+++ my-app/src/pages/Home/index.jsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react';
+import { useSearchParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { onAuthStateChanged } from 'firebase/auth';
 import { UserIdState } from '../../atom/authRecoil';
@@ -11,11 +12,22 @@ import DessertIcon from '../../assets/Icon-dessert.png';
 import CategoryCard from '../../components/home/CategoryCard';
 import RecentPosts from '../../components/home/RecentPosts';
 import usePost from '../../hooks/usePost';
+import ToastMessage from '../../components/notification/ToastMessage';
 
 function Home() {
   const userId = useRecoilValue(UserIdState);
   const [userName, setUserName] = useState('');
   const { isLoading, isError, data: posts } = usePost(userId, 'ALL');
+  const [searchParams, setSearchParams] = useSearchParams();
+  const [successToast, setSuccessToast] = useState(false);
+
+  useEffect(() => {
+    const getSuccessToast = searchParams.get('success');
+
+    if (getSuccessToast) {
+      activeToast(true);
+    }
+  }, [location]);
 
   useEffect(() => {
     const unsubscribe = onAuthStateChanged(appAuth, (user) => {
@@ -31,6 +43,18 @@ function Home() {
 
   if (isError) {
     return <div>fetch data중 에러</div>;
+  }
+
+  function activeToast(isSuccess) {
+    setSuccessToast(isSuccess);
+    const timer = setTimeout(() => {
+      setSuccessToast(false);
+      setSearchParams(false);
+    }, 2000);
+
+    return () => {
+      clearTimeout(timer);
+    };
   }
 
   const postCount = posts.length;
@@ -91,6 +115,7 @@ function Home() {
         </section>
       </S.Container>
       <NavBar />
+      {successToast && <ToastMessage message={'오늘의 커디어리 등록 완료!'} />}
     </>
   );
 }
